refactor(ci): move webui jobs in a dedicated folder
