+++ .circleci/ci/src/jobs/frontend/index.ts
@@ -1,0 +1,1 @@
+export * from './job-webui-build';
