+++ .circleci/ci/src/jobs/index.ts
@@ -16,4 +16,3 @@ export * from './job-setup';
 export * from './job-slack-announcement';
 export * from './job-sonarcloud-analysis';
 export * from './job-validate';
-export * from './job-webui-build';
